+++ FinancialBloom/Backend/spring-backend/BackEnd/src/main/java/csc450/BackEnd/BudgetRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BudgetRepository extends JpaRepository<Budget, Integer> {
+public interface BudgetRepository extends JpaRepository<Budget, Long> {
 
     // Optional custom queries:
     Budget findByName(String name);
